Move the standard library to a std crate
This commit splits the standard library contents (all commands and
functions) out of the core crate and puts them in their own std crate.

This is to fulfill the design principle of keeping a tiny core for the
language, as this allows us to drop almost all dependencies from the
core crate.
