+++ .github/workflows/package.sh
@@ -24,5 +24,6 @@ set -eux
 if grep 'Changes in' NEWS.md | head -n 1 | fgrep 'X.Y.Z'; then
     echo "Skipping endbasic publish test in development version"
 else
+    ( cd std && cargo publish --dry-run )
     ( cd cli && cargo publish --dry-run )
 fi
