+++ core/src/lib.rs
@@ -21,48 +21,10 @@
 #![warn(unused, unused_extern_crates, unused_import_braces, unused_qualifications)]
 #![warn(unsafe_code)]
 
-use crate::exec::MachineBuilder;
-use std::cell::RefCell;
-use std::rc::Rc;
-
 // TODO(jmmv): Should narrow the exposed interface by 1.0.0.
 pub mod ast;
-pub mod console;
-mod editor;
 pub mod eval;
 pub mod exec;
-pub mod help;
 mod lexer;
-pub mod numerics;
 mod parser;
 mod reader;
-pub mod store;
-pub mod strings;
-
-/// Creates a new machine builder populated with all scripting commands from the standard library.
-pub fn scripting_machine_builder(console: Rc<RefCell<dyn console::Console>>) -> MachineBuilder {
-    let mut builder = MachineBuilder::default();
-
-    builder = console::add_all(builder, console);
-    builder = exec::add_all(builder);
-    builder = numerics::add_all(builder);
-    builder = strings::add_all(builder);
-
-    builder
-}
-
-/// Creates a new machine builder populated with all scripting _and_ interactive commands from the
-/// standard library.
-pub fn interactive_machine_builder(
-    console: Rc<RefCell<dyn console::Console>>,
-    store: Rc<RefCell<dyn store::Store>>,
-) -> MachineBuilder {
-    let mut builder = scripting_machine_builder(console.clone());
-
-    let program = Rc::from(RefCell::from(editor::Editor::default()));
-
-    builder = help::add_all(builder, console.clone());
-    builder = store::add_all(builder, program, console, store);
-
-    builder
-}
