+++ std/Cargo.toml
@@ -1,37 +1,32 @@
 [package]
-name = "endbasic"
+name = "endbasic-std"
 version = "0.4.0"
 license = "Apache-2.0"
 authors = ["Julio Merino <julio@meroh.net>"]
 categories = ["development-tools", "parser-implementations"]
 keywords = ["basic", "interpreter", "learning", "programming"]
-description = "The EndBASIC programming language - CLI"
+description = "The EndBASIC programming language - standard library"
 homepage = "https://github.com/jmmv/endbasic"
 repository = "https://github.com/jmmv/endbasic"
 readme = "README.md"
 edition = "2018"
 
-[features]
-default = ["bin"]
-bin = ["crossterm"]
-
-[[bin]]
-name = "endbasic"
-required-features = ["bin"]
-
 [dependencies]
-anyhow = "1.0"
 async-trait = "0.1"
-crossterm = { version = "0.18", optional = true }
-dirs = "2.0"
 endbasic-core = { version = "0.4.0", path = "../core" }
-endbasic-std = { version = "0.4.0", path = "../std" }
-futures-lite = "1.1"
-getopts = "0.2"
-thiserror = "1.0"
 time = { version = "0.2", features = ["std"] }
 
+# We don't directly use getrandom but rand does, and we have to customize how
+# getrandom is built if we want it to work in a WASM context.
+[dependencies.getrandom]
+version = "0.2"
+features = ["js"]
+
+[dependencies.rand]
+version = "0.8"
+default-features = false
+features = ["small_rng", "std"]
+
 [dev-dependencies]
-filetime = "0.2"
+futures-lite = "1.1"
 regex = "1"
-tempfile = "3"
