+++ std/NOTICE
@@ -1,0 +1,2 @@
+EndBASIC
+Copyright 2020-2021 Julio Merino
