+++ std/tests/interactive.bas
