+++ std/tests/script.bas
