+++ std/tests/script.err
