+++ std/tests/script.out
