+++ web/Cargo.toml
@@ -23,7 +23,7 @@ default = ["console_error_panic_hook"]
 async-channel = "1.5"
 async-trait = "0.1"
 console_error_panic_hook = { version = "0.1.6", optional = true }
-endbasic-core = { version = "0.4.0", path = "../core" }
+endbasic-std = { version = "0.4.0", path = "../std" }
 js-sys = "0.3"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
